Moved the mapping function into initializers
