+++ main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	. "google-scraper/helpers"
 	_ "google-scraper/initializers"
 	_ "google-scraper/routers"
 
@@ -9,8 +8,5 @@ import (
 )
 
 func main() {
-	web.AddFuncMap("displayStatus", DisplayStatus)
-	web.AddFuncMap("displayFormattedCreatedDate", DisplayFormattedCreatedDate)
-
 	web.Run()
 }
